requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.34.0
sqlalchemy>=2.0.45
psycopg2-binary>=2.9.11
pydantic[email]>=2.10.4
email-validator>=2.3.0
pydantic-settings>=2.12.0
python-jose[cryptography]>=3.5.0
bcrypt==4.0.1
passlib[bcrypt]==1.7.4
python-multipart>=0.0.21
python-dotenv>=1.2.1
aiohttp>=3.9.1
requests>=2.32.5
qrcode>=8.2
Pillow>=12.1.0
# Google Sheets API
google-api-python-client>=2.100.0
google-auth-httplib2>=0.1.1
google-auth-oauthlib>=1.0.0
# Testing
pytest>=7.4.3
httpx>=0.25.1

alembic>=1.13.1
razorpay>=1.3.0
